+++ config_provider.py
@@ -29,8 +29,7 @@ class ConfigProvider:
     def get_powermeter_min_point(self):
         """
         The minimum power of your powermeter for the normal "regulation loop".
-        if your powermeter jumps under this point, the limit will be reduced instantly. it is like a "super high priority limit change".
-        if you defined ON_GRID_FEED_JUMP_TO_LIMIT_PERCENT > 0, then the limit will jump to the defined percent when it is lower than this point.
+        if your powermeter jumps under this point and ON_GRID_FEED_FAST_LIMIT_DECREASE = true, the limit will be reduced instantly. it is like a "super high priority limit change".
         """
         pass
 
@@ -40,9 +39,9 @@ class ConfigProvider:
         """
         pass
     
-    def on_grid_feed_jump_to_limit_percent(self):
+    def on_grid_feed_fast_limit_decrease(self):
         """
-        If the powermeter falls below the min point, the limit will be increased to this percent of the powermeter value.
+        If the powermeter falls below the min point, the limit will be immediatelly decreased.
         """
         pass    
 
@@ -109,8 +108,8 @@ class ConfigFileConfigProvider(ConfigProvider):
     def on_grid_usage_jump_to_limit_percent(self):
         return self.config.getint('COMMON', 'ON_GRID_USAGE_JUMP_TO_LIMIT_PERCENT')
     
-    def on_grid_feed_jump_to_limit_percent(self):
-        return self.config.getint('COMMON', 'ON_GRID_FEED_JUMP_TO_LIMIT_PERCENT')    
+    def on_grid_feed_fast_limit_decrease(self):
+        return self.config.getboolean('COMMON', 'ON_GRID_FEED_FAST_LIMIT_DECREASE')    
 
     def get_min_wattage_in_percent(self, inverter_idx):
         return self.config.getint('INVERTER_' + str(inverter_idx + 1), 'HOY_MIN_WATT_IN_PERCENT')
@@ -172,8 +171,10 @@ class OverridingConfigProvider(ConfigProvider):
             else:
                 logger.error(f"Unknown inverter key {key}")
         else:
-            if key in ['powermeter_target_point', 'powermeter_max_point', 'powermeter_min_point', 'powermeter_tolerance', 'on_grid_usage_jump_to_limit_percent', 'on_grid_feed_jump_to_limit_percent']:
+            if key in ['powermeter_target_point', 'powermeter_max_point', 'powermeter_min_point', 'powermeter_tolerance', 'on_grid_usage_jump_to_limit_percent']:
                 return int(value)
+            elif key in ['on_grid_feed_fast_limit_decrease']:
+                return bool(value)
             else:
                 logger.error(f"Unknown common key {key}")
 
@@ -214,8 +215,8 @@ class OverridingConfigProvider(ConfigProvider):
     def on_grid_usage_jump_to_limit_percent(self):
         return self.common_config.get('on_grid_usage_jump_to_limit_percent')
     
-    def on_grid_feed_jump_to_limit_percent(self):
-        return self.common_config.get('on_grid_feed_jump_to_limit_percent')    
+    def on_grid_feed_fast_limit_decrease(self):
+        return self.common_config.get('on_grid_feed_fast_limit_decrease')    
 
     def get_min_wattage_in_percent(self, inverter_idx):
         if inverter_idx >= len(self.inverter_config):
@@ -255,7 +256,7 @@ class MqttConfigProvider(OverridingConfigProvider):
         self.min_point = None
         self.tolerance = None
         self.on_grid_usage_jump_to_limit_percent = None
-        self.on_grid_feed_jump_to_limit_percent = None
+        self.on_grid_feed_fast_limit_decrease = None
         self.min_wattage_in_percent = []
         self.normal_wattage = []
         self.reduce_wattage = []